requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-crispy-forms==1.14.0
django-htmx==1.8.0
Pillow==9.0.1
sqlparse==0.4.2
